Split off SuperconductingQubitNoiseProperties
