+++ cirq-core/cirq/protocols/json_test_data/SuperconductingQubitsNoiseProperties.json
@@ -1,5 +1,5 @@
 {
-  "cirq_type": "NoiseProperties",
+  "cirq_type": "SuperconductingQubitsNoiseProperties",
   "gate_times_ns": [
     [
       "CZPowGate",
